+++ CHANGELOG.md
@@ -1,3 +1,25 @@
+# Changelog for RESTPresenter
+
+## [0.2.3](https://github.com/xtend-packages/rest-presenter/compare/0.2.2...0.2.3) (2024-03-27)
+
+
+### Bug Fixes
+
+* lunarphp/core move to dev dependency ([1c14c78](https://github.com/xtend-packages/rest-presenter/commit/1c14c78b42f232cec812cbf7240beb0b1413aecb))
+* Make sure custom resources directory is created with setup ([724a347](https://github.com/xtend-packages/rest-presenter/commit/724a347148941a4675627f5b8db42c7e3fa79094))
+
+
+### Code Refactoring
+
+* Remove api routes register via XtendRouter with AutoDiscover ([19a0b40](https://github.com/xtend-packages/rest-presenter/commit/19a0b4019c82614bf9099da91a639bf4912e6ee0))
+
+## 0.2.2 (2024-03-27)
+
+
+### Code Refactoring
+
+* XtendRoute more control over routes with better IDE support ([f0843cc](https://github.com/xtend-packages/rest-presenter/commit/f0843cc837394f460405d91822834dd2e058a8fd))
+
 # Changelog for REST Presenter
 
 ## 0.2.1 (26-03-2024)